Renamed 'McForge' to 'MCForge'. We need to maintain consistency.
It is also pronounced 'Emm See Forge'. There is no potayto-potahto thing going on here.

Potahto is not a word.

Also renamed Main.cs to frmMain.cs to reduce confusion later on.

We need to decide how we are going to case our variables and methods. We can differentiate between public and private methods / variables if we want but consistency and organization need to be a high concern.